@@ -2,27 +2,24 @@
 using Microsoft.VisualStudio;
 using Microsoft.VisualStudio.Shell;
 using NuGetMonitor.Services;
-using System;
 using System.Runtime.InteropServices;
-using System.Threading;
 using Task = System.Threading.Tasks.Task;
 
-namespace NuGetMonitor
+namespace NuGetMonitor;
+
+[Guid(PackageGuidString)]
+[PackageRegistration(UseManagedResourcesOnly = true, AllowsBackgroundLoading = true)]
+[ProvideAutoLoad(VSConstants.UICONTEXT.SolutionExistsAndFullyLoaded_string, PackageAutoLoadFlags.BackgroundLoad)]
+public sealed class NuGetMonitorPackage : ToolkitPackage
 {
-    [Guid(PackageGuidString)]
-    [PackageRegistration(UseManagedResourcesOnly = true, AllowsBackgroundLoading = true)]
-    [ProvideAutoLoad(VSConstants.UICONTEXT.SolutionExistsAndFullyLoaded_string, PackageAutoLoadFlags.BackgroundLoad)]
-    public sealed class NuGetMonitorPackage : ToolkitPackage
+    public const string PackageGuidString = "38279e01-6b27-4a29-9221-c4ea8748f16e";
+
+    protected override async Task InitializeAsync(CancellationToken cancellationToken, IProgress<ServiceProgressData> progress)
     {
-        public const string PackageGuidString = "38279e01-6b27-4a29-9221-c4ea8748f16e";
+        await JoinableTaskFactory.SwitchToMainThreadAsync(cancellationToken);
 
-        protected override async Task InitializeAsync(CancellationToken cancellationToken, IProgress<ServiceProgressData> progress)
-        {
-            await JoinableTaskFactory.SwitchToMainThreadAsync(cancellationToken);
+        MonitorService.RegisterEventHandler();
 
-            MonitorService.RegisterEventHandler();
-
-            MonitorService.CheckForUpdates().FireAndForget();
-        }
+        MonitorService.CheckForUpdates().FireAndForget();
     }
 }
