@@ -1,30 +1,28 @@
 ﻿using Community.VisualStudio.Toolkit;
 using Microsoft.VisualStudio.Shell;
-using System.Threading.Tasks;
 
-namespace NuGetMonitor.Services
+namespace NuGetMonitor.Services;
+
+public static class MonitorService
 {
-    public static class MonitorService
+    public static void RegisterEventHandler()
     {
-        public static void RegisterEventHandler()
-        {
-            VS.Events.SolutionEvents.OnAfterOpenSolution += SolutionEvents_OnAfterOpenSolution;
-            VS.Events.SolutionEvents.OnAfterCloseSolution += SolutionEvents_OnAfterCloseSolution;
-        }
+        VS.Events.SolutionEvents.OnAfterOpenSolution += SolutionEvents_OnAfterOpenSolution;
+        VS.Events.SolutionEvents.OnAfterCloseSolution += SolutionEvents_OnAfterCloseSolution;
+    }
 
-        private static void SolutionEvents_OnAfterCloseSolution()
-            => InfoBarService.CloseInfoBar();
+    private static void SolutionEvents_OnAfterCloseSolution()
+        => InfoBarService.CloseInfoBar();
 
-        private static void SolutionEvents_OnAfterOpenSolution(Solution solution)
-            => CheckForUpdates().FireAndForget();
+    private static void SolutionEvents_OnAfterOpenSolution(Solution? solution)
+        => CheckForUpdates().FireAndForget();
 
-        public static async Task CheckForUpdates()
-        {
-            var packageIdentities = await ProjectService.GetPackageReferences().ConfigureAwait(true);
+    public static async Task CheckForUpdates()
+    {
+        var packageIdentities = await ProjectService.GetPackageReferences().ConfigureAwait(true);
 
-            var packageReferences = await NuGetService.CheckPackageReferences(packageIdentities).ConfigureAwait(true);
+        var packageReferences = await NuGetService.CheckPackageReferences(packageIdentities).ConfigureAwait(true);
 
-            await InfoBarService.ShowInfoBar(packageReferences).ConfigureAwait(true);
-        }
+        await InfoBarService.ShowInfoBar(packageReferences.ToArray()).ConfigureAwait(true);
     }
 }
